refactor(userspace): extract event-stream wire format to protocol_events.go
Eleventh and final code-motion increment of the protocol.go wire-protocol
split (#4839). Moves the binary event-stream framing (helper -> daemon
push stream) constants and the event-stream status type into a cohesive
sibling protocol_events.go in the same userspace package.

Moved verbatim (no rename, no logic or json-tag change): EventStreamStatus,
EventFrameHeaderSize, the event-message-type constant block, and the
session-event flag-bit constant block.

Behavior byte-identical. With this increment protocol.go is reduced to
the control envelope (ControlRequest/ControlResponse), the ConfigSnapshot
root and its remaining cross-cutting sub-snapshots, and the version
constant block. Build passes; gofmt clean.
